quickstart/requirements.txt
bentoml
torch
transformers
a2wsgi                                       1.10.7
accelerate                                   1.2.1
aiohappyeyeballs                             2.4.4
aiohttp                                      3.11.10
aiosignal                                    1.3.2
aiosqlite                                    0.20.0
annotated-types                              0.7.0
anyio                                        4.7.0
appdirs                                      1.4.4
asgiref                                      3.8.1
attrs                                        24.3.0
beautifulsoup4                               4.12.3
bentoml                                      1.3.16
cattrs                                       23.1.2
cbor                                         1.0.0
certifi                                      2024.12.14
charset-normalizer                           3.4.0
circus                                       0.18.0
click                                        8.1.7
click-option-group                           0.5.6
cloudpickle                                  3.1.0
colorama                                     0.4.6
datasets                                     2.19.0
deepmerge                                    2.0
Deprecated                                   1.2.15
dill                                         0.3.8
filelock                                     3.16.1
FlagEmbedding                                1.3.3
frozenlist                                   1.5.0
fs                                           2.4.16
fsspec                                       2024.3.1
grpcio                                       1.68.1
grpcio-tools                                 1.68.1
h11                                          0.14.0
httpcore                                     1.0.7
httpx                                        0.28.1
httpx-ws                                     0.7.0
huggingface-hub                              0.27.0
idna                                         3.10
ijson                                        3.3.0
importlib-metadata                           6.11.0
inflection                                   0.5.1
inscriptis                                   2.5.0
ir_datasets                                  0.5.9
Jinja2                                       3.1.4
joblib                                       1.4.2
lxml                                         5.3.0
lz4                                          4.3.3
markdown-it-py                               3.0.0
MarkupSafe                                   3.0.2
mdurl                                        0.1.2
mpmath                                       1.3.0
multidict                                    6.1.0
multiprocess                                 0.70.16
networkx                                     3.4.2
numpy                                        2.2.0
nvidia-ml-py                                 12.560.30
opentelemetry-api                            1.20.0
opentelemetry-instrumentation                0.41b0
opentelemetry-instrumentation-aiohttp-client 0.41b0
opentelemetry-instrumentation-asgi           0.41b0
opentelemetry-sdk                            1.20.0
opentelemetry-semantic-conventions           0.41b0
opentelemetry-util-http                      0.41b0
packaging                                    24.2
pandas                                       2.2.3
pathspec                                     0.12.1
peft                                         0.14.0
pillow                                       11.0.0
pip                                          22.3.1
pip-requirements-parser                      32.0.1
pipeline                                     0.1.0
prometheus_client                            0.21.1
prompt-toolkit                               3.0.36
propcache                                    0.2.1
protobuf                                     5.29.1
psutil                                       6.1.0
pyarrow                                      18.1.0
pyarrow-hotfix                               0.6
pydantic                                     2.10.3
pydantic_core                                2.27.1
Pygments                                     2.18.0
pyparsing                                    3.2.0
python-dateutil                              2.9.0.post0
python-dotenv                                1.0.1
python-json-logger                           3.2.1
python-multipart                             0.0.20
pytz                                         2024.2
PyYAML                                       6.0.2
pyzmq                                        26.2.0
questionary                                  2.0.1
regex                                        2024.11.6
requests                                     2.32.3
rich                                         13.9.4
safetensors                                  0.4.5
schema                                       0.7.7
scikit-learn                                 1.6.0
scipy                                        1.14.1
sentence-transformers                        3.3.1
sentencepiece                                0.2.0
setuptools                                   65.5.1
simple-di                                    0.1.5
six                                          1.17.0
sniffio                                      1.3.1
soupsieve                                    2.6
starlette                                    0.42.0
sympy                                        1.13.1
yarl                                         1.18.3
zipp                                         3.21.0
zlib-state                                   0.1.9
